rename codegen submodule.
