+++ src/bloqade/compiler/codegen/hardware/__init__.py
@@ -1,0 +1,11 @@
+from .lattice import GenerateLattice
+from .lattice_site_coefficients import GenerateLatticeSiteCoefficients
+from .piecewise_constant import GeneratePiecewiseConstantChannel
+from .piecewise_linear import GeneratePiecewiseLinearChannel
+
+__all__ = [
+    "GenerateLattice",
+    "GenerateLatticeSiteCoefficients",
+    "GeneratePiecewiseConstantChannel",
+    "GeneratePiecewiseLinearChannel",
+]
